enterprise-qa-data/requirements.txt
pyyaml
python-dotenv
rank-bm25
nltk
pytest